isolate position tests.
After introducing the keepInViewport method on the Popup object
the tests with positioning started failing in PhantomJS. Because
of its headless nature you should only do really basic UI level
tests in it. This test failed because the rendered DOM had a
smaller than anticipated viewport.

I exposed the Popup object on the _internals property on the
bindingHandler. This was the least intrusive way to make it
testable.

The tests is functionally equivalent.

To fix the test for repositioning I had to stub out the
keepInViewport method to an identify function. Poor mans
isolation...