use jquery event handling directly instead of ko.utils.registerEventHandler
We need to use jquery's event handling (which is what
ko.utils.registerEventHandler does when jquery is available)
to be able to catch events triggered by jquery#trigger.

We already depend on jquery being available so we might as
well also use jquery for setting up a listener on window.

Using ko.utils.registerEventHandler the handler will not be
freed until the node is disposed, which never happens when we
listen on window. So the existing solution caused a leak of
event handlers. Since we use the same instance of the PopupTemplate
class it ended up adding a new subscription for each PopupTemplate
that you opened.

Using jquery's event handler means that we can remove the
listener in our cleanup hook, which was not possible using the
knockout equivalent, as it would do so itself when removing the
node.