add one esc key handler and close the latest popup
This came up when nesting popup templates. If you relied on the
close on escape behavior, and had a popup within a popup, both of
them would be closed when you pressed escape.

Actually, the parent would be closed, which would then go to clean
up any nested popups, but the effect is the same.

To fix this, I register a single event listener for the escape
press, and only close the most recently opened popup.

For all typical usages the result will be the same, but it allows
us to nest them without attempting to reimplement the closeOnEsc
behavior in the consuming project.